fix character capsule size issue.
In project Foxtrot, the capsule of the fox should be in fact a lying down cylinder.
Because its an quadruped animal.

So in fact in the code the height is the length of the fox, and the radius
is a half of the height of the fox.

Also, the radius of the capsule need to be multiplied by the scale_y.